+++ tox/src/main.rs
@@ -17,8 +17,8 @@ use std::fs::File;
 use std::io::Read;
 use std::rc::Rc;
 use structopt::StructOpt;
-use syntax::lexer::Lexer;
-// use syntax::parser::Parser;
+use syntax::parser::Parser;
+
 use util::emmiter::Reporter;
 use util::symbol::{SymbolFactory, Symbols};
 use vm::VM;
@@ -61,8 +61,8 @@ pub fn run_interpreter(path: String, ptokens: bool, past: bool) {
 
     let strings = Rc::new(SymbolFactory::new());
     let mut symbols = Symbols::new(Rc::clone(&strings));
-
-    let ast = match ::syntax::compiler::Parser::new(input, reporter.clone(), &mut symbols).parse() {
+    
+    let ast = match Parser::new(input, reporter.clone(), &mut symbols).parse() {
         Ok(statements) => statements,
         Err(_) => {
             reporter.emit(input);
@@ -71,7 +71,7 @@ pub fn run_interpreter(path: String, ptokens: bool, past: bool) {
     };
 
     if past {
-    println!("{:#?}", ast);
+        println!("{:#?}", ast);
     }
 
     let mut infer = Infer::new();
@@ -125,10 +125,9 @@ pub fn run(path: String, ptokens: bool, past: bool) {
     let strings = Rc::new(SymbolFactory::new());
     let mut symbols = Symbols::new(Rc::clone(&strings));
 
-    let ast = match ::syntax::compiler::Parser::new(input, reporter.clone(), &mut symbols).parse() {
+    let ast = match Parser::new(input, reporter.clone(), &mut symbols).parse() {
         Ok(statements) => statements,
         Err(_) => {
-            
             reporter.emit(input);
             ::std::process::exit(65)
         }
